[refactor] 네이버 로그인 코드 간소화 #85
