+++ app/src/main/java/com/readme/android/navigator/MainNavigatorImpl.kt
@@ -3,6 +3,7 @@ package com.readme.android.navigator
 import android.content.Context
 import com.readme.android.MainActivity
 import com.readme.android.auth.ui.LoginActivity
+import com.readme.android.core_ui.ext.startActivity
 import com.readme.android.set_nick_name.SetNickNameActivity
 import com.readme.android.write_feed.FeedWriteActivity
 import javax.inject.Inject
